# frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "ShapeShift",
      "description": "OpenAPI Schema for ShapeShift",
      "version": "0.0.1"
    },
    "paths": {
...
      "/api/v1/user/create": {
        "post": {
          "tags": ["users"],
          "summary": "Create a new user",
          "operationId": "createUser",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAuth"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/UserOut"
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request - User with this email or username already exists",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            },
            "500": {
              "description": "Internal Server Error - An unexpected error occurred",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/user/me": {
        "get": {
          "tags": ["users"],
          "summary": "Get current user",
          "operationId": "getCurrentUser",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/UserOut"
                  }
                }
              }
            }
          },
          "security": [
            {
              "JWT": []
            }
          ]
        }
      },
      "/api/v1/user/update": {
        "put": {
          "tags": ["users"],
          "summary": "Update current user",
          "operationId": "updateCurrentUser",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserUpdate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/UserOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          },
          "security": [
            {
              "JWT": []
            }
          ]
        }
      },
      "/api/v1/user/delete": {
        "delete": {
          "tags": ["users"],
          "summary": "Delete current user",
          "operationId": "deleteCurrentUser",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {}
                }
              }
            }
          },
          "security": [
            {
              "JWT": []
            }
          ]
        }
      },
...
      "/api/v1/workout/": {
        "get": {
          "tags": ["workouts"],
          "summary": "Get all workouts",
          "operationId": "getWorkouts",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/WorkoutOut"
                    },
                    "type": "array",
                    "title": "Response Getworkouts"
                  }
                }
              }
            }
          },
          "security": [
            {
              "JWT": []
            }
          ]
        },
        "post": {
          "tags": ["workouts"],
          "summary": "Create a workout",
          "operationId": "createWorkout",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Workout"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          },
          "security": [
            {
              "JWT": []
            }
          ]
        }
      },
      "/api/v1/workout/{workout_id}": {
        "get": {
          "tags": ["workouts"],
          "summary": "Get a single workout",
          "operationId": "getWorkout",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "workout_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Workout Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/WorkoutOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": ["workouts"],
          "summary": "Update a workout",
          "operationId": "updateWorkout",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "workout_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Workout Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutUpdate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/WorkoutOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": ["workouts"],
          "summary": "Delete a workout",
          "operationId": "deleteWorkout",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "workout_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Workout Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {}
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/v1/session/workout/{workout_id}": {
        "get": {
          "tags": ["sessions"],
          "summary": "Get all sessions for a workout",
          "operationId": "getWorkoutSessions",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "workout_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Workout Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/SessionWorkoutOut"
                    },
                    "title": "Response Getworkoutsessions"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "ErrorResponse": {
          "properties": {
            "detail": {
              "type": "string",
              "title": "Detail"
            }
          },
          "type": "object",
          "required": ["detail"],
          "title": "ErrorResponse"
        },
...
        "Gender": {
          "type": "string",
          "enum": ["male", "female", "other"],
          "title": "Gender"
        },
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "SessionWorkoutOut": {
          "properties": {
            "session_id": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            },
            "date": {
              "type": "string",
              "format": "date-time",
              "title": "Date"
            },
            "duration_mins": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Duration Mins"
            },
            "active": {
              "type": "boolean",
              "title": "Active"
            }
          },
          "type": "object",
          "required": ["session_id", "date", "duration_mins", "active"],
          "title": "SessionWorkoutOut"
        },
...
        "UserAuth": {
          "properties": {
            "email": {
              "type": "string",
              "title": "Email",
              "description": "user email"
            },
            "username": {
              "type": "string",
              "maxLength": 50,
              "minLength": 5,
              "title": "Username",
              "description": "user username"
            },
            "password": {
              "type": "string",
              "maxLength": 50,
              "minLength": 5,
              "title": "Password",
              "description": "user password"
            }
          },
          "type": "object",
          "required": ["email", "username", "password"],
          "title": "UserAuth"
        },
        "UserOut": {
          "properties": {
            "user_id": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            },
            "username": {
              "type": "string",
              "title": "Username"
            },
            "email": {
              "type": "string",
              "title": "Email"
            },
            "first_name": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "First Name"
            },
            "last_name": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last Name"
            },
            "birth_date": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Birth Date"
            },
            "gender": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Gender"
                },
                {
                  "type": "null"
                }
              ]
            },
            "height_in": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Height In"
            },
            "weight": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Weight"
            }
          },
          "type": "object",
          "required": [
            "user_id",
            "username",
            "email",
            "first_name",
            "last_name",
            "birth_date",
            "gender",
            "height_in",
            "weight"
          ],
          "title": "UserOut"
        },
        "UserUpdate": {
          "properties": {
            "email": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Email"
            },
            "first_name": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "First Name"
            },
            "last_name": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last Name"
            },
            "birth_date": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Birth Date"
            },
            "gender": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Gender"
                },
                {
                  "type": "null"
                }
              ]
            },
            "height_in": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Height In"
            },
            "weight": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Weight"
            }
          },
          "type": "object",
          "required": [
            "email",
            "first_name",
            "last_name",
            "birth_date",
            "gender",
            "height_in",
            "weight"
          ],
          "title": "UserUpdate"
        },
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": ["loc", "msg", "type"],
          "title": "ValidationError"
        },
        "Workout": {
          "properties": {
            "_id": {
              "anyOf": [
                {
                  "type": "string",
                  "example": "5eb7cf5a86d9755df3a6c593"
                },
                {
                  "type": "null"
                }
              ],
              "title": " Id",
              "description": "MongoDB document ObjectID"
            },
            "workout_id": {
              "type": "string",
              "format": "uuid",
              "title": "Workout Id",
              "unique": true
            },
            "title": {
              "type": "string",
              "title": "Title"
            },
            "created_date": {
              "type": "string",
              "format": "date-time",
              "title": "Created Date"
            },
            "duration_mins": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Duration Mins"
            },
            "day_of_week": {
              "items": {
                "type": "integer"
              },
              "type": "array",
              "title": "Day Of Week",
              "default": []
            },
            "owner": {
              "properties": {
                "id": {
                  "type": "string",
                  "title": "Id"
                },
                "collection": {
                  "type": "string",
                  "title": "Collection"
                }
              },
              "type": "object",
              "required": ["id", "collection"],
              "title": "Owner"
            }
          },
          "type": "object",
          "required": ["title", "owner"],
          "title": "Workout"
        },
        "WorkoutCreate": {
          "properties": {
            "title": {
              "type": "string",
              "maxLength": 50,
              "minLength": 1,
              "title": "Title"
            },
            "day_of_week": {
              "items": {
                "type": "integer"
              },
              "type": "array",
              "title": "Day Of Week",
              "default": []
            }
          },
          "type": "object",
          "required": ["title"],
          "title": "WorkoutCreate"
        },
        "WorkoutOut": {
          "properties": {
            "workout_id": {
              "type": "string",
              "format": "uuid",
              "title": "Workout Id"
            },
            "title": {
              "type": "string",
              "title": "Title"
            },
            "created_date": {
              "type": "string",
              "format": "date-time",
              "title": "Created Date"
            },
            "duration_mins": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Duration Mins"
            },
            "day_of_week": {
              "items": {
                "type": "integer"
              },
              "type": "array",
              "title": "Day Of Week",
              "default": []
            }
          },
          "type": "object",
          "required": ["workout_id", "title", "created_date", "duration_mins"],
          "title": "WorkoutOut"
        },
        "WorkoutUpdate": {
          "properties": {
            "title": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 50,
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Title"
            },
            "duration_mins": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Duration Mins"
            },
            "day_of_week": {
              "items": {
                "type": "integer"
              },
              "type": "array",
              "title": "Day Of Week",
              "default": []
            }
          },
          "type": "object",
          "required": ["title", "duration_mins"],
          "title": "WorkoutUpdate"
        }
      },
      "securitySchemes": {
        "JWT": {
          "type": "oauth2",
          "flows": {
            "password": {
              "scopes": {},
              "tokenUrl": "/api/v1/auth/login"
            }
          }
        }
      }
    },
    "servers": [
      {
        "url": "https://shapeshift-backend.onrender.com"
      }
    ]
  }